backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]>=2.8
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.10
tenacity>=8.2.3
boto3==1.28.39
aiosmtplib
email-validator
bcrypt==4.1.3
passlib[bcrypt]==1.7.4